Moved entire stats dir from Logic to Model
