+++ src/main/java/seedu/address/model/Model.java
@@ -9,7 +9,7 @@ import javafx.beans.value.ObservableBooleanValue;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.battle.Battle;
-import seedu.address.logic.statistics.PlayerStatistics;
+import seedu.address.model.statistics.PlayerStatistics;
 import seedu.address.model.battleship.Battleship;
 import seedu.address.model.battleship.Orientation;
 import seedu.address.model.cell.Cell;
