+++ src/main/java/seedu/address/model/ModelManager.java
@@ -21,7 +21,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.battle.Battle;
 import seedu.address.logic.battle.BattleManager;
-import seedu.address.logic.statistics.PlayerStatistics;
+import seedu.address.model.statistics.PlayerStatistics;
 import seedu.address.model.battleship.Battleship;
 import seedu.address.model.battleship.Orientation;
 import seedu.address.model.cell.Cell;
