+++ src/main/java/seedu/address/model/statistics/PlayerStatistics.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.statistics;
+package seedu.address.model.statistics;
 
 
 /**
